backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.6
pydantic==2.6.1
pydantic-settings==2.1.0
ultralytics==8.1.14
torch==2.2.0
torchvision==0.17.0
